价格相关的字段改为float

+++ model/albumOrder.go
@@ -20,9 +20,9 @@ import (
 type Albumorder struct {
 	ID           uint                        `gorm:"primary_key" gqlschema:"delete!;query!;querys" description:"订单id"`
 	UserId       uint                        `gorm:"DEFAULT:0;NOT NULL;" gqlschema:"create!;querys" description:"创建用户id" funservice:"employee"`
-	SinglePrice  uint                        `gorm:"DEFAULT:0;NOT NULL;" gqlschema:"create;querys" description:"单价"`
+	SinglePrice  float64                     `gorm:"DEFAULT:0;NOT NULL;" gqlschema:"create;querys" description:"单价"`
 	Amount       uint                        `gorm:"DEFAULT:0;NOT NULL;" gqlschema:"create;querys" description:"数量"`
-	TotalPrice   uint                        `gorm:"DEFAULT:0;NOT NULL;" gqlschema:"create;querys" description:"总价"`
+	TotalPrice   float64                     `gorm:"DEFAULT:0;NOT NULL;" gqlschema:"create;querys" description:"总价"`
 	Specs        AlbumOrderSpecsEnumType     `gorm:"DEFAULT:1;NOT NULL;" gqlschema:"create;querys" description:"纸张规格"`
 	Material     AlbumOrderMaterialEnumType  `gorm:"DEFAULT:1;NOT NULL;" gqlschema:"create;querys" description:"材质"`
 	Template     string                      `gorm:"Type:varchar(255);DEFAULT:'';NOT NULL;" gqlschema:"create;querys" description:"使用的相册模板"`
@@ -32,11 +32,11 @@ type Albumorder struct {
 	CreatedAt    time.Time                   `description:"创建时间" gqlschema:"querys"`
 	UpdatedAt    time.Time                   `description:"更新时间" gqlschema:"querys"`
 	DeletedAt    *time.Time
-	v2           int    `gorm:"-" exclude:"true"`
-	Remark       string `gorm:"-" exclude:"true" gqlschema:"create" description:"订单备注"`
-	Address      string `gorm:"-" exclude:"true" gqlschema:"create" description:"收货地址"`
-	FreightPrice uint   `gorm:"-" exclude:"true" gqlschema:"create" description:"运费"`
-	OrderStatus  string `gorm:"-" exclude:"true" gqlschema:"create!" description:"订单状态:待支付:TO_BE_PAID,待发货:TO_BE_DELIVER"`
+	v2           int     `gorm:"-" exclude:"true"`
+	Remark       string  `gorm:"-" exclude:"true" gqlschema:"create" description:"订单备注"`
+	Address      string  `gorm:"-" exclude:"true" gqlschema:"create" description:"收货地址"`
+	FreightPrice float64 `gorm:"-" exclude:"true" gqlschema:"create" description:"运费"`
+	OrderStatus  string  `gorm:"-" exclude:"true" gqlschema:"create!" description:"订单状态:待支付:TO_BE_PAID,待发货:TO_BE_DELIVER"`
 }
 
 type Albumorders struct {
@@ -65,8 +65,8 @@ type OrderActionProp struct {
 }
 
 // 跨接口创建订单 childrenType = "albumorder"
-func createOrder(userId uint, childrenId uint, childrenType string, remark string, address string, status string, freightPrice uint, goodsPrice uint) (result interface{}, err error) {
-	mutation := `mutation ($address: String, $childrenId: Int!, $childrenType: String!, $freightPrice: Int, $goodsPrice: Int, $remark: String, $status: OrderStatusEnumType!, $userId: Int!) {
+func createOrder(userId uint, childrenId uint, childrenType string, remark string, address string, status string, freightPrice float64, goodsPrice float64) (result interface{}, err error) {
+	mutation := `mutation ($address: String, $childrenId: Int!, $childrenType: String!, $freightPrice: Float, $goodsPrice: Float, $remark: String, $status: OrderStatusEnumType!, $userId: Int!) {
 				  order {
 					orderinfos {
 					  action {
@@ -139,7 +139,7 @@ func (o Albumorder) Create(params graphql.ResolveParams) (Albumorder, error) {
 	o.OrderStatus = p["orderStatus"].(string)
 	// template
 	if p["singlePrice"] != nil {
-		o.SinglePrice = uint(p["singlePrice"].(int))
+		o.SinglePrice = p["singlePrice"].(float64)
 	}
 	if p["proId"] != nil {
 		fmt.Println(reflect.TypeOf(p["proId"]))
@@ -168,7 +168,7 @@ func (o Albumorder) Create(params graphql.ResolveParams) (Albumorder, error) {
 		o.Amount = uint(p["amount"].(int))
 	}
 	if p["totalPrice"] != nil {
-		o.TotalPrice = uint(p["totalPrice"].(int))
+		o.TotalPrice = p["totalPrice"].(float64)
 	}
 	if p["specs"] != nil {
 		o.Specs = p["specs"].(AlbumOrderSpecsEnumType)
@@ -189,7 +189,7 @@ func (o Albumorder) Create(params graphql.ResolveParams) (Albumorder, error) {
 		o.Address = p["address"].(string)
 	}
 	if p["freightPrice"] != nil {
-		o.FreightPrice = uint(p["freightPrice"].(int))
+		o.FreightPrice = p["freightPrice"].(float64)
 	}
 	// TODO 事务
 	err := db.Create(&o).Error
